% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bioness.R
\name{read_bioness}
\alias{read_bioness}
\title{Read BIONESS electronic files
Returns a BIONESS class list object with metadata represented as a list and
data represented as a dataframe.}
\usage{
read_bioness(file)
}
\arguments{
\item{file}{}
}
\value{
a BIONESS class object with a /{metadata} and /{data} section
}
\description{
Read BIONESS electronic files
Returns a BIONESS class list object with metadata represented as a list and
data represented as a dataframe.
}
